+++ CHANGELOG.md
@@ -91,6 +91,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
     - some conflict resolution scenarios when upgrading several packages
     - not handling conflicting files errors during the installation process
     - some environment variables are not available during the common operations (install, upgrade, downgrade, uninstall, makepkg, launch)
+    - not displaying and uninstalling dependent packages during conflict resolutions
+    - not retrieving all packages that would break if a given package is uninstalled
     - AUR
         - info dialog of installed packages displays the latest PKGBUILD file instead of the one used for installation/upgrade/downgrade (the fix will only work for new installed packages)
         - multi-threaded download: not retrieving correctly some source files URLs (e.g: linux-xanmod-lts)